use specific tag version of refstack for installation.
It's not convenient for development to deploy latest master.
For example we need to commit features in several commits
and we can break something in the middle.
Also we want to implement something by several developers
and in this case we can have invalid code in the middle.

So we need to deploy refstack website with specific
and released (fully implemented and tested) version.
When next version will be released then we up version here.

Change-Id: Ie4459d1d5d3415670237aaeabfd56c8b62da275d