Ensure pymysql is installed
PyMySQL was/will be removed from requirements.txt in refstack as it
is not an explicit dependency. It's an optional DB driver that is only
necessary if it is specified in the refstack configuration database URL.
Since pymysql is specified as the driver here, it is necessary to ensure
it is installed.

Change-Id: I336aa8ef196526601c16b165d8200b63ff11825d