Switch to the nodejs puppet module
As we do with ether{calc,pad} and other NodeJS-based services, use the
nodejs puppet module to set up nodejs and npm instead of relying on the
distro-provided packages.

This changes the apt repository source from the default Ubuntu
repository to deb.nodesource.com and also updates the NodeJS version to
the latest 4.x version available from nodesource.

Change-Id: Ia3223e3b802edb55135632ec1969c6b0d0a3391c